--- a/run_perf_tests.sh
+++ b/run_perf_tests.sh
@@ -2,5 +2,6 @@
 dotnet restore
 
 for path in test/*.PerformanceTests/*.csproj; do
+    dotnet test -f netcoreapp2.1 -c Release ${path}
     dotnet test -f netcoreapp3.1 -c Release ${path}
 done